Process intervals in blocks, instead of points within interval in blocks
The previous implementation used a grid of `(num_intervals,)`, where
each program would process all of the points in an interval, blockwise,
in parallel. This is optimal if there are many points in an interval.
However, in some cases, we don't have many points in an interval, so its
actually better to process the intervals blockwise with each program
processing a block of the intervals.